Add maintenance mode and support for decimal prices in the bot
Introduces maintenance mode toggle for admins, enables decimal prices for products, updates stock display to include duration, adds a purchase report channel, fixes image navigation issues, and improves the trusted sellers list. Includes a migration script for price column type change.

+++ bot/models/product.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, Boolean, ForeignKey
+from sqlalchemy import Column, Integer, String, Text, Boolean, ForeignKey, Float
 from sqlalchemy.orm import relationship
 from .base import Base, TimestampMixin
 
@@ -26,7 +26,7 @@ class ProductPrice(Base, TimestampMixin):
     id = Column(Integer, primary_key=True, autoincrement=True)
     product_id = Column(Integer, ForeignKey("products.id"), nullable=False)
     duration = Column(String(100), nullable=False)
-    price = Column(Integer, nullable=False)
+    price = Column(Float, nullable=False)
     
     product = relationship("Product", back_populates="prices")
